// src/app/models/question/question.model.ts
export class Question {
  ans?: string;
  ans1?: string;
  ans2?: string;
  ans3?: string;
  ans4?: string;
  ansid?: number;
  correct_ans?: string;
  qfield?: string;
  qid?: null;
  qu?: string;
  stuid?: null;
  subid?: number;
}

export class PostQuestion {
  subid?: number;
  question?: string;
  ans1?: string;
  ans2?: string;
  ans3?: string;
  ans4?: string;
  correct_ans?: string;
}